backend/requirements.txt
asgiref==3.11.1
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.12.1
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
